moved all azure service principal to storage credential migration code to databricks.labs.ucx.azure from databricks.labs.ucx.migration
