+++ src/databricks/labs/ucx/azure/azure_credentials.py
