+++ src/databricks/labs/ucx/cli.py
@@ -19,7 +19,7 @@ from databricks.labs.ucx.hive_metastore import ExternalLocations, TablesCrawler
 from databricks.labs.ucx.hive_metastore.mapping import TableMapping
 from databricks.labs.ucx.hive_metastore.table_migrate import TableMove, TablesMigrate
 from databricks.labs.ucx.install import WorkspaceInstallation
-from databricks.labs.ucx.migration.azure_credentials import (
+from databricks.labs.ucx.azure.azure_credentials import (
     AzureServicePrincipalMigration,
 )
 from databricks.labs.ucx.workspace_access.groups import GroupManager
